Sistema de classificação feito

+++ src/database/migrations/20200428230631-create-Anuncio.js
@@ -38,7 +38,7 @@ module.exports = {
     },
 
     valor:{
-      type:Sequelize.INTEGER,
+      type:Sequelize.DECIMAL(5,2),
       allowNull:false
     },
